Moved Server class to own file
Moved Server class to own file
