+++ src/ClientHandler.java
@@ -7,48 +7,6 @@ import java.text.*;
 import java.util.*; 
 import java.net.*; 
   
-// Server class 
-public class Server  
-{ 
-    public static void main(String[] args) throws IOException  
-    { 
-        // server is listening on port 5056 
-        ServerSocket ss = new ServerSocket(5056); 
-          
-        // running infinite loop for getting 
-        // client request 
-        while (true)  
-        { 
-            Socket s = null; 
-              
-            try 
-            { 
-                // socket object to receive incoming client requests 
-                s = ss.accept(); 
-                  
-                System.out.println("A new client is connected : " + s); 
-                  
-                // obtaining input and out streams 
-                DataInputStream dis = new DataInputStream(s.getInputStream()); 
-                DataOutputStream dos = new DataOutputStream(s.getOutputStream()); 
-                  
-                System.out.println("Assigning new thread for this client"); 
-  
-                // create a new thread object 
-                Thread t = new ClientHandler(s, dis, dos); 
-  
-                // Invoking the start() method 
-                t.start(); 
-                  
-            } 
-            catch (Exception e){ 
-                s.close(); 
-                e.printStackTrace(); 
-            } 
-        } 
-    } 
-} 
-  
 // ClientHandler class 
 class ClientHandler extends Thread  
 { 
@@ -83,7 +41,7 @@ class ClientHandler extends Thread
                 // receive the answer from client 
                 received = dis.readUTF(); 
                   
-                if(received.equals("Exit")) 
+                if(received.equalsIgnoreCase("Exit")) 
                 {  
                     System.out.println("Client " + this.s + " sends exit..."); 
                     System.out.println("Closing this connection.");
